screenshoter: set anroid and ios tests to run on sauce, and bumped concurrency up +1
For context, the Sauce account is currently limited to 4 concurrent
sessions. If another team is running tests then tests will start to
queue and lead to timeout errors. The ios and android tests run in
series (first android and then ios) so bumping concurrency to 2
(from 1) should be an OK blend of speeding up these suites, while also
allowing some room in the queue for others.